+++ src/components/sources/index.ts
@@ -13,7 +13,6 @@ import OlSourceTileJSON from "./OlSourceTileJSON.vue";
 import OlSourceTileWMS from "./OlSourceTileWMS.vue";
 import OlSourceVector from "./OlSourceVector.vue";
 import OlSourceVectorTile from "./OlSourceVectorTile.vue";
-import OlSourceWebglVector from "./OlSourceWebglVector.vue";
 import OlSourceXYZ from "./OlSourceXYZ.vue";
 import OlSourceWMTS from "./OlSourceWMTS.vue";
 
@@ -32,7 +31,6 @@ function install(app: App) {
   app.component("ol-source-tile-wms", OlSourceTileWMS);
   app.component("ol-source-vector", OlSourceVector);
   app.component("ol-source-vector-tile", OlSourceVectorTile);
-  app.component("ol-source-webglvector", OlSourceWebglVector);
   app.component("ol-source-xyz", OlSourceXYZ);
   app.component("ol-source-wmts", OlSourceWMTS);
 }
@@ -55,7 +53,6 @@ export {
   OlSourceTileWMS,
   OlSourceVector,
   OlSourceVectorTile,
-  OlSourceWebglVector,
   OlSourceXYZ,
   OlSourceWMTS,
 };